+++ tools/tslint-rules/noViewEncapsulationRule.ts
@@ -1,6 +1,7 @@
-const Lint = require('tslint');
-const path = require('path');
-const minimatch = require('minimatch');
+import * as path from 'path';
+import * as ts from 'typescript';
+import * as Lint from 'tslint';
+import * as minimatch from 'minimatch';
 
 const ERROR_MESSAGE = 'Components must turn off view encapsulation.';
 
@@ -10,28 +11,34 @@ const ERROR_MESSAGE = 'Components must turn off view encapsulation.';
  * Rule that enforces that view encapsulation is turned off on all components.
  * Files can be whitelisted via `"no-view-encapsulation": [true, "\.spec\.ts$"]`.
  */
-class Rule extends Lint.Rules.AbstractRule {
-  apply(file) {
-    return this.applyWithWalker(new Walker(file, this.getOptions()));
+export class Rule extends Lint.Rules.AbstractRule {
+  apply(sourceFile: ts.SourceFile) {
+    return this.applyWithWalker(new Walker(sourceFile, this.getOptions()));
   }
 }
 
 class Walker extends Lint.RuleWalker {
-  constructor(file, options) {
-    super(...arguments);
+
+  /** Whether the walker should check the current source file. */
+  private _enabled: boolean;
+
+  constructor(sourceFile: ts.SourceFile, options: Lint.IOptions) {
+    super(sourceFile, options);
 
     // Globs that are used to determine which files to lint.
     const fileGlobs = options.ruleArguments || [];
 
     // Relative path for the current TypeScript source file.
-    const relativeFilePath = path.relative(process.cwd(), file.fileName);
+    const relativeFilePath = path.relative(process.cwd(), sourceFile.fileName);
 
     // Whether the file should be checked at all.
     this._enabled = fileGlobs.some(p => minimatch(relativeFilePath, p));
   }
 
   visitClassDeclaration(node) {
-    if (!this._enabled || !node.decorators) return;
+    if (!this._enabled || !node.decorators) {
+      return;
+    }
 
     node.decorators
       .map(decorator => decorator.expression)
@@ -50,5 +57,3 @@ class Walker extends Lint.RuleWalker {
   }
 
 }
-
-exports.Rule = Rule;
